split out mock swarm
