+++ monad-crypto/src/secp256k1.rs
@@ -4,7 +4,7 @@ use sha2::Digest;
 #[derive(Copy, Clone, Debug, PartialOrd, Ord, PartialEq, Eq, Hash)]
 pub struct PubKey(secp256k1::PublicKey);
 pub struct KeyPair(secp256k1::KeyPair);
-#[derive(Copy, Clone, Debug)]
+#[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
 pub struct Signature(secp256k1::ecdsa::RecoverableSignature);
 
 #[derive(Debug)]
